Wrote test code and data analysis code
For the test code: Took chunks of code from testing individual components and made it so I could collect the data. Added interrupts to make sure data was collected often enough and at the right time. Made different files for different ways of turning power on and off to the SMA. Manual triggering is not updated, but just takes an amount of time from the used and turns power on to the SMA for that amount of time. I have code to turn on power to the SMA periodically and turn off power to the SMA once it reaches a certain encoder position. I have code to turn on power to the SMA once it reaches a certain encoder position and turn off power once it reaches another encoder position.
For the data processing code: Initial, code takes data from a .txt file and reads and interprets the data. The code tells the efficiency and outputs a graph showing the power, position, and force with respect to time.