Cleaned up code and fixed Manual Trigger
Manual Trigger code was just messed up, I think because of just weird things happening in the Arduino IDE and files not syncing or something. Fixed that. Also removed all of the "debug code" from the rest of the code so essentially none of the comments are just commenting out code.